Created Transaction java class. This class is used for to deposit, withdraw and transfer funds between accounts.
Modified Customer and Account classes to include customerExists() and accountExists() functions to determine whether a customer or account exists in the database.

Referenced the functions above in all functions to ensure that a corresponding customer/account exists before updates or queries are made to the database.

Changed MySQL Database to have customerID, accountID, and transactionID all auto-increment. This removes the need for getting a random value in Customer,Account, and Transaction classes.