Changed how all the functions were connecting to the DB. Created DBConnection.java class to connect to the MySQL DB. Handles select, insert, update, and delete. These functions are called by passing the parameters: Table, Where, Values, and Parameters, ensuring reuse with all DB Tables.
Changed Account.java, Customer.java, and Transaction.java classes to use the new DBConnection Class to connect to DB and run select, insert, update, and delete functions.

Updated Account and Customer classes to act as a record from the DB upon initialization.

Account and Customer classes have 2 constructors. The first is for an existing Account/Customer (which sets values of the object equal to the values found in the DB). The second is for a new Account/Customer that needs to be inserted into the DB.